requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytesseract==0.3.13
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
